chore: commonises shared types and merge functionality
