+++ apps/manage/src/app/views/s62a/cases/view/questions.ts
@@ -82,13 +82,9 @@ export function getQuestions(answers: S62aCaseViewModel, isQuestionView?: boolea
 		validators: [new CILAmountValidator(), new CILAmountLengthValidator()]
 	};
 
-	// TODO: Make sure to remove these error expectations.
-	// @ts-expect-error - the view model update is coming in next PR
 	const isIndividual = answers?.applicantType === APPLICANT_TYPE_ID.INDIVIDUAL;
-	// @ts-expect-error - the view model update is coming in next PR
 	const manageApplicantOrganisations = !isIndividual ? (answers?.manageApplicantOrganisations as ApplicantOrg[]) : [];
 	const applicantOrganisationOptions = getApplicantOrganisationOptions(manageApplicantOrganisations);
-	// @ts-expect-error - the view model update is coming in next PR
 	const hasAgent = answers?.hasAgent === BOOLEAN_OPTIONS.YES;
 
 	const applicantContactsValidator = getApplicantContactsValidator(hasAgent, isIndividual);
@@ -1202,7 +1198,9 @@ export function getQuestions(answers: S62aCaseViewModel, isQuestionView?: boolea
 			emptyListText: 'No applicants found',
 			showAnswersInSummary: true,
 			emptyStateAddStyle: 'prominent',
-			maximumAnswers: 10
+			maximumAnswers: 10,
+			removalPrompt:
+				'Removing this organisation will also remove any linked contacts. You will not be able to undo this.'
 		},
 		applicantOrganisationName: {
 			type: COMPONENT_TYPES.SINGLE_LINE_INPUT,
